Fix for undersized hole on A Drive Frame Lower
The hole for the double bearing stack on the Lower part of the A Drive Frame was undersized compared to other holes for the pin mod.
Removed chamfer, pulled the hole from 1.40mm to 1.65mm to create the correct sized hole, then re-added the 0.4mm chamfer.
STL & CAD updated, CAD also has A Drive Frame Upper bodies renamed to better reflect the various umbilical mount sizings.